Fix protected members being erased when unwrapping store refs
When a vue store is passed through `reactive` or returned from a `ref`/`computed`, the type can get broken if it has protected members. The vue wrap/unwrap process uses a mapped type, which effectively erases the non-public members.

To fix that we mark all the stores as `ShallowReactive`, which causes the Vue wrap/unwrap process to pass them through untouched.